# backend/docs/swagger.yaml
definitions:
...
  bid.CreateBidDTO:
    properties:
      amount:
        type: integer
      auction_id:
        type: integer
    required:
    - amount
    - auction_id
    type: object
  bid.RetrieveBidDTO:
    properties:
      amount:
        type: integer
      auction_id:
        type: integer
      bidder_id:
        type: integer
    required:
    - amount
    - auction_id
    - bidder_id
    type: object
...
  custom_errors.HTTPError:
    properties:
      error_description:
        type: string
    type: object
...
  manufacturer.RetrieveManufacturerDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.RetrieveModelDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: BD2
  description: Car-Dealer API
  license:
    name: MIT
  title: Car‑Dealer API
  version: "1.0"
paths:
...
  /bid:
    get:
      consumes:
      - application/json
      description: Retrieve all bids
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/bid.RetrieveBidDTO'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get all bids
      tags:
      - bid
    post:
      consumes:
      - application/json
      description: Create a new bid for an auction
      parameters:
      - description: Bid details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bid.CreateBidDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created bid
          schema:
            $ref: '#/definitions/bid.RetrieveBidDTO'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a new bid
      tags:
      - bid
  /bid/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a bid by its ID
      parameters:
      - description: Bid ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bid details
          schema:
            $ref: '#/definitions/bid.RetrieveBidDTO'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get bid by ID
      tags:
      - bid
  /bid/auction/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all bids placed on a specific auction
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/bid.RetrieveBidDTO'
            type: array
        "400":
          description: Invalid auction ID or retrieval error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get bids by auction ID
      tags:
      - bid
  /bid/bidder/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all bids placed by a specific bidder
      parameters:
      - description: Bidder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/bid.RetrieveBidDTO'
            type: array
        "400":
          description: Invalid bidder ID or retrieval error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get bids by bidder ID
      tags:
      - bid
  /bid/highest/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves the highest bid for a specific auction
      parameters:
      - description: Auction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Highest bid details
          schema:
            $ref: '#/definitions/bid.RetrieveBidDTO'
        "400":
          description: Invalid auction ID or retrieval error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get the highest bid for an auction
      tags:
      - bid
  /bid/highest/auction/{auctionID}/bidder/{bidderID}:
    get:
      consumes:
      - application/json
      description: Retrieves the highest bid placed by a specific user on a specific
        auction
      parameters:
      - description: Auction ID
        in: path
        name: auctionID
        required: true
        type: integer
      - description: Bidder ID
        in: path
        name: bidderID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Highest bid details
          schema:
            $ref: '#/definitions/bid.RetrieveBidDTO'
        "400":
          description: Invalid auction ID, bidder ID, or retrieval error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get the highest bid by a user for a specific auction
      tags:
      - bid
...
  /car/manufacturers:
    get:
      consumes:
      - application/json
      description: Returns a list of all manufacturers stored in the database.
      produces:
      - application/json
      responses:
        "200":
          description: List of manufacturers
          schema:
            items:
              $ref: '#/definitions/manufacturer.RetrieveManufacturerDTO'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get all manufacturers
      tags:
      - car
  /car/models/id/{id}:
    get:
      consumes:
      - application/json
      description: Returns a list of all models stored in the database for a given
        manufacturer id.
      parameters:
      - description: Manufacturer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of models
          schema:
            items:
              $ref: '#/definitions/model.RetrieveModelDTO'
            type: array
        "400":
          description: ID is not a number
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "404":
          description: Models not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get all models by manufacturer id
      tags:
      - car
  /car/models/name/{name}:
    get:
      consumes:
      - application/json
      description: Returns a list of all models stored in the database for a given
        manufacturer name.
      parameters:
      - description: Manufacturer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of models
          schema:
            items:
              $ref: '#/definitions/model.RetrieveModelDTO'
            type: array
        "404":
          description: Models not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      summary: Get all models by manufacturer name
      tags:
      - car
...
  /favourite/dislike/{id}:
    delete:
      consumes:
      - application/json
      description: Dislike offer by giving it's id. You have to be logged in to perform
        this operation.
      parameters:
      - description: Sale offer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized - user not logged in
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "404":
          description: Sale offer not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - Bearer: []
      summary: Dislike offer
      tags:
      - favourite
  /favourite/like/{id}:
    post:
      consumes:
      - application/json
      description: Like new offer by giving it's id. You have to be logged in to perform
        this operation.
      parameters:
      - description: Sale offer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Liked offer
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "401":
          description: Unauthorized - user not logged in
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "404":
          description: Sale offer not found
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/custom_errors.HTTPError'
      security:
      - Bearer: []
      summary: Like new offer
      tags:
      - favourite
...
schemes:
- http
swagger: "2.0"